Coding_Agent/requirements.txt
## requirements.txt

flask==2.0.1
requests==2.26.0
pytest==6.2.5
python-dotenv==0.19.1
gitpython==3.1.24